templating the test
